backend/requirements.txt
fastapi
uvicorn
requests
vaderSentiment
python-dotenv
transformers
torch
numpy
pandas
scikit-learn
textblob
nltk 